Fix mapping of GraphQl Int type to Java int type instead of Long (#901)
Closes https://github.com/apollographql/apollo-android/issues/896

+++ apollo-compiler/src/test/graphql/com/example/arguments_complex/TestQuery.java
@@ -50,7 +50,7 @@ public final class TestQuery implements Query<TestQuery.Data, Optional<TestQuery
 
   private final TestQuery.Variables variables;
 
-  public TestQuery(@Nonnull Input<Episode> episode, long stars, double greenValue) {
+  public TestQuery(@Nonnull Input<Episode> episode, int stars, double greenValue) {
     Utils.checkNotNull(episode, "episode == null");
     variables = new TestQuery.Variables(episode, stars, greenValue);
   }
@@ -92,7 +92,7 @@ public final class TestQuery implements Query<TestQuery.Data, Optional<TestQuery
   public static final class Builder {
     private Input<Episode> episode = Input.absent();
 
-    private long stars;
+    private int stars;
 
     private double greenValue;
 
@@ -104,7 +104,7 @@ public final class TestQuery implements Query<TestQuery.Data, Optional<TestQuery
       return this;
     }
 
-    public Builder stars(long stars) {
+    public Builder stars(int stars) {
       this.stars = stars;
       return this;
     }
@@ -127,13 +127,13 @@ public final class TestQuery implements Query<TestQuery.Data, Optional<TestQuery
   public static final class Variables extends Operation.Variables {
     private final Input<Episode> episode;
 
-    private final long stars;
+    private final int stars;
 
     private final double greenValue;
 
     private final transient Map<String, Object> valueMap = new LinkedHashMap<>();
 
-    Variables(Input<Episode> episode, long stars, double greenValue) {
+    Variables(Input<Episode> episode, int stars, double greenValue) {
       this.episode = episode;
       this.stars = stars;
       this.greenValue = greenValue;
@@ -148,7 +148,7 @@ public final class TestQuery implements Query<TestQuery.Data, Optional<TestQuery
       return episode;
     }
 
-    public long stars() {
+    public int stars() {
       return stars;
     }
 
@@ -169,7 +169,7 @@ public final class TestQuery implements Query<TestQuery.Data, Optional<TestQuery
           if (episode.defined) {
             writer.writeString("episode", episode.value != null ? episode.value.rawValue() : null);
           }
-          writer.writeLong("stars", stars);
+          writer.writeInt("stars", stars);
           writer.writeDouble("greenValue", greenValue);
         }
       };
